All Files Needed to start distance routine.
Includes borrowed drive and sonar classes. Lift currently has an issue where if not stopped before next movement command is activated, the "drift" can cause it to desync due to the way the encoder currently works. I'll need to talk to dr. Picorillo to see if I can't get that second cable wired in so that I can tell direction from the encoder too.